aggiunto controlli sicurezza per piatto il ristorante vede solo i suoi piatti, non può modificare o vedere altri piatti di altri user

+++ app/Http/Controllers/Admin/PlateController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Plate;
+use App\Restaurant;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Auth;
@@ -18,7 +19,8 @@ class PlateController extends Controller
      */
     public function index()
     {
-        $plates = Plate::all();
+        $user = Auth::user();
+        $plates = $user->restaurant->plates;
         return view('admin.plate.index', compact('plates'));
     }
 
@@ -50,9 +52,11 @@ class PlateController extends Controller
         $data = $request->all();
         $newPlate = new Plate();
         $newPlate->fill($data);
-        $newPlate->slug = $this->getSlug($data['name']);
-        $newPlate->user_id = Auth::user()->id;
         $newPlate->is_visible = isset($newPlate->is_visible);
+
+        // associo il piatto al ristorante
+        $newPlate->restaurant_id = Auth::id();
+
         $newPlate->save();
         //reindirizzo a un altra pagina
         return redirect()->route('admin.plate.index');
@@ -66,6 +70,11 @@ class PlateController extends Controller
      */
     public function show(Plate $plate)
     {
+        // controllo per impedire modifica di piatti di altri ristoranti
+        if($plate->restaurant_id !== Auth::id()){
+            abort(403);  //403 per mostrare un errore di permessi
+        }
+
         return view('admin.plate.show', compact('plate'));
     }
 
@@ -76,7 +85,11 @@ class PlateController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit(Plate $plate)
-    {
+    {   
+        // controllo per impedire modifica di piatti di altri ristoranti
+        if($plate->restaurant_id !== Auth::id()){
+            abort(403);  //403 per mostrare un errore di permessi
+        }
         return view('admin.plate.edit', compact('plate'));
     }
 
@@ -89,6 +102,11 @@ class PlateController extends Controller
      */
     public function update(Request $request, Plate $plate)
     {
+        // controllo per impedire modifica di piatti di altri ristoranti
+        if($plate->restaurant_id !== Auth::id()){
+            abort(403);  //403 per mostrare un errore di permessi
+        }
+        
         //validazione
         $request->validate([
             'name' => 'required | string | max:50'
@@ -110,6 +128,11 @@ class PlateController extends Controller
      */
     public function destroy(Plate $plate)
     {
+        // controllo per impedire modifica di piatti di altri ristoranti
+        if($plate->restaurant_id !== Auth::id()){
+            abort(403);  //403 per mostrare un errore di permessi
+        }
+        
         $plate->delete();
         return redirect()->route('admin.plate.index');
     }
